+++ telegram_messages_dump/exceptions/__init__.py
@@ -1,0 +1,2 @@
+from .DumpingError  import DumpingError
+from .MetadataError import MetadataError
