+++ telegram_messages_dump/settings/__init__.py
@@ -1,0 +1,1 @@
+from .ChatDumpSettings import ChatDumpSettings
